pyproject.toml
[project]
name = "broadcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sylvio Campos Neto", email = "sylvio.campos@warren.com.br" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "truststore>=0.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.10",
]
